Fix import stat scripts

+++ app/models/updaters/pve/base_stats.rb
@@ -5,15 +5,15 @@ class Updaters::Pve::BaseStats
 
   def update
     @users.each do |user|
-      user.pve_base_stat.destroy
+      user.pve_base_stat.destroy if user.pve_base_stat
       ::PveBaseStat.create({
                             user: user,
-                            kills: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['kills']['basic']['value'],
-                            deaths: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['deaths']['basic']['value'],
-                            kd: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['killsDeathsRatio']['basic']['value'],
-                            suicides: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['suicides']['basic']['value'],
-                            assists: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['assists']['basic']['value'],
-                            most_kills: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['bestSingleGameKills']['basic']['value']
+                            kills: @stats[user.id]['mergedAllCharacters']['results']['allPvE']['allTime']['kills']['basic']['value'],
+                            deaths: @stats[user.id]['mergedAllCharacters']['results']['allPvE']['allTime']['deaths']['basic']['value'],
+                            kd: @stats[user.id]['mergedAllCharacters']['results']['allPvE']['allTime']['killsDeathsRatio']['basic']['value'],
+                            suicides: @stats[user.id]['mergedAllCharacters']['results']['allPvE']['allTime']['suicides']['basic']['value'],
+                            assists: @stats[user.id]['mergedAllCharacters']['results']['allPvE']['allTime']['assists']['basic']['value'],
+                            most_kills: @stats[user.id]['mergedAllCharacters']['results']['allPvE']['allTime']['bestSingleGameKills']['basic']['value']
                           })
     end
   end
