+++ app/models/updaters/pvp/base_stats.rb
@@ -5,7 +5,7 @@ class Updaters::Pvp::BaseStats
 
   def update
     @users.each do |user|
-      user.pvp_base_stat.destroy
+      user.pvp_base_stat.destroy if user.pvp_base_stat
       ::PvpBaseStat.create({
                             user: user,
                             kills: @stats[user.id]['mergedAllCharacters']['results']['allPvP']['allTime']['kills']['basic']['value'],
